requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
geoalchemy2==0.15.0
shapely==2.0.3